+++ packages/language-server/src/launcher.ts
@@ -1,4 +1,5 @@
 import { handleComposerNamespaceCommand } from "./composer-namespace.js";
+import { runRenameCommand } from "./rename-command.js";
 
 if (process.argv[2] === "--infer-composer-namespace") {
   void handleComposerNamespaceCommand([{ directoryUri: process.argv[3] }])
@@ -6,6 +7,8 @@ if (process.argv[2] === "--infer-composer-namespace") {
     .catch(() => {
       process.exitCode = 1;
     });
+} else if (process.argv[2] === "--rename") {
+  void runRenameCommand().then((response) => process.stdout.write(`${JSON.stringify(response)}\n`));
 } else {
   void import("./node.js");
 }